Rename all 'zoom-pan' to 'zoom-and-pan' in file names
