+++ packages/dx-chart-core/src/plugins/zoom-and-pan/computeds.ts
