+++ packages/dx-chart-core/src/types/plugins.zoom-and-pan.types.ts
