+++ packages/dx-react-chart-bootstrap4/src/plugins/zoom-and-pan.jsx
@@ -1,5 +1,5 @@
 import { ZoomAndPan as ZoomAndPanBase } from '@devexpress/dx-react-chart';
 import { withComponents } from '@devexpress/dx-react-core';
-import { DragBox } from '../templates/zoom-pan/drag-box';
+import { DragBox } from '../templates/zoom-and-pan/drag-box';
 
 export const ZoomAndPan = withComponents({ DragBox })(ZoomAndPanBase);
