+++ packages/dx-react-chart-bootstrap4/src/templates/zoom-and-pan/drag-box.jsx
