+++ packages/dx-react-chart-material-ui/src/templates/zoom-and-pan/drag-box.jsx
