+++ packages/dx-react-chart/src/index.ts
@@ -17,7 +17,7 @@ export { SplineSeries } from './plugins/spline-series';
 export { Axis, ArgumentAxis, ValueAxis } from './plugins/axis';
 export { Tooltip } from './plugins/tooltip';
 
-export * from './plugins/zoom-pan';
+export * from './plugins/zoom-and-pan';
 
 export * from './plugins/event-tracker';
 export * from './plugins/hover-state';
