+++ packages/dx-react-chart/src/plugins/zoom-and-pan.tsx
